Bug 1992138: Check ownership of cluster and infraenv in rhsso auth (#2471)
* Bug 1992138: Check ownership of cluster and infraenv in rhsso auth

This commit changes the rhsso authenticator to check for either the
cluster or infra-env owner depending on what is present in the
parameters.

Previously we never checked for the infra env owner. This effectively
implements authz for infra-env.

This commit also copies the cluster owner information to the infra-env
when it is created. This was required to get the new subsystem tests to
pass.

https://bugzilla.redhat.com/show_bug.cgi?id=1992138

* Bug 1992138: Copy existing cluster user info to infra env

This commit adds a migration to copy the user information used for authz
from existing clusters to their corresponding infra env if one exists
and if that infra env does not already have user information set.

Without this, any existing installs that have not yet finished would
start to fail because requests from the agent against the infra env
would fail authz checks. In this case the user would have to create a
new cluster/infraenv to get things working again.

https://bugzilla.redhat.com/show_bug.cgi?id=1992138