Bug 1992138: Handle possible NULLs in user data migration (#2488)
There are cases where the user data in either the cluster or infra env
could be NULL in the database. This fails the migration with the
following error:

`sql: Scan error on column index 1, name \"user_name\": converting NULL
to string is unsupported`

This commit handles these cases by using the sql.NullString type in the
Scan call which is specifically made to handle a value from a database
that can either be NULL or a valid string.

https://bugzilla.redhat.com/show_bug.cgi?id=1992138