MGMT-6670: Fix BelongsToMachineCidr validator message
This PR changes the message returned by the BelongsToMachineCidr
validator. The previous assumption that there is always a single machine
network and this one should appear in the messages is no longer valid.

Given that passing the state of the validators around happens via
passing boolean statuses, we decide to drop the specific value from the
message and keep it as simple as "host belongs to all" or "host does not
belong to some", with a pointer to where to look for in case of an
error.

Follows-up: #2414
Contributes-to: MGMT-6670