MGMT-6670: Fix handling of empty network in various network params
This PR fixes handling of Machine, Cluster and Service Networks when
the value passed is a non-empty array with empty objects.

The current assumption that the MachineNetworks (or other) variable is
either empty or contains a valid set of CIDRs is incorrect, as it has
been discovered that the following requests arrive to the service

```
--data '{"machine_networks": [{}]}'
```

Those should be rather arriving as

```
--data '{"machine_networks": []}'
```

but in order to handle all the possibilities, we add a scenario for the
former in this change.

Contributes-to: MGMT-6670