MGMT-7593: Added custom manifest feature usage
In order to know that a user has requested to add a custom manifest, for metrics counting for user usage (and not internal usage), We needed to change all internal manifests creations calls to use the internal function directly instead of calling the API handler.

Currently we only add a usage on create and we dont remove it on delete. This is because we do need to discuss this flow. If a user override a built in manifest and deletes it etc.

The manifests manager has the ability to remove a feature usage with the new `setUsage` function - just not yet called from the delete API.