Bug 2000612: Use the agent's inventory hostname as the source for BMH's hostname
The hostname was not being set in the BMH's hardware details when the
hostname annotation was not used. This is because the reconcile step was
using agent.Spec.Hostname as the source of hostname rather than the
inventory. The former is latter set (after inspection) while the former
is only set when the annotation is set

Signed-off-by: Flavio Percoco <flavio@redhat.com>