MGMT-7498: Add missing backwards compatibility for network configuration
This commit adds a backwards compatibility for setting network
configuration using the following structures

* MachineNetworkCidr
* ClusterNetworkCidr
* ClusterNetworkHostPrefix
* ServiceNetworkCidr

This works in a way that whenever the old data structure is received by
the service, the respective values in both old and new structures are
set internally.

In case when service receives both old and new data structure in the
same request, the old data structure will take precedence.

Follows-up: openshift#2455
Contributes-to: MGMT-7498
Contributes-to: Bug 1999752