MGMT-7736: Do not blindly load the whole database in InfraEnv migration (#2555)
This PR changes the behaviour of the 20210713123129_populate_infra_env
database migration. What happens currently is that the migration is
using the application code (namely `common.GetClustersFromDBWhere`) and
preloads the following tables without checking whether they are actually
needed

```
var ClusterSubTables = [...]string
{
  HostsTable,
  MonitoredOperatorsTable,
  ClusterNetworksTable,
  ServiceNetworksTable,
  MachineNetworksTable
}
```

This is not a correct behaviour for those reasons

* the application code should not be used inside the migration itself,
  as the migrations should be agnostic to the version that executes it
* only the data actually needed should be loaded, as the initial state
  of the DB at the moment of running the migration may not be the same
  as the initial state of the DB at the moment of writing the code of
  the migration

Please note this PR is a bugfix for enabling the upgrade from ocm-2.3
version to the current master. The correct fix should include replacing
the `common.GetClustersFromDBWhere` call with `tx.Model().Select()`, but
this is outside of the scope of this commit.

Co-authored-by: Mat Kowalski <mko@redhat.com>