MGMT-7742: Fixing host getting stuck in "binding" state (#2561)
Once Bind command isssued, the host moves to Binding state and its ClusterID is set
to the target cluster. So, from instruction manager POV this host is now bound host and
not unbound host. The fix is to move binding instuction from unboind host instruction manager
commands set, to the bound host instruction manager command set.
In addition fixing the security definitions for Bind (adn unbind) command. It should be user
auth, and not agent auth