MGMT-7253: Add assisted-image-service to AgentServiceconfig (#2480)
* MGMT-7253: asc should deploy assisted-image-service

Update the agentserviceconfig controller to deploy the
assisted-image-service. Include the assisted-image-service image
references in the operator manifests.

* NO-ISSUE: Move ports to IntOrString type

It gets painful trying to manage if the port should be int, int32,
string, or IntOrString. Instead of having to deal with all of those by
starting with an int32 constance, we will maintain our ports as
IntOrString vars that we can transform as neeeded.

* MGMT-7253: Provide RHCOS_VERSIONS to image-service

Take the OS_IMAGES that is being provided to the assisted-service and
also give it to the assisted-image-service.

* NO-ISSUE: use urlAuth on infraenv file download

* NO-ISSUE: provide ca bundle to ais

* MGMT-7253: Update tests for assisted-image-service

Update the agentserviceconfig's controller tests to verify we handle the
different assisted-image-service components.

* NO-ISSUE: Use less official looking volume mounts