MGMT-7315: Allow providing machine networks in all cases
This commit adds an ability to manually provide machine networks. This
is in order to enable the following use cases

* dual-stack clusters
* remote worker nodes

Previously whenever VIPs were provided, machine network was
automatically calculated based on them and the available networks on the
hosts. This is not a valid case anymore

* in case of a dual-stack cluster, as multiple VIPs are not allowed
* in case of remote worker nodes, as nodes do not belong to the same L2
  as VIPs

In those scenarios we want to enable users to explicitly provide machine
networks so that the clusters can be installed.

A sample CR presenting dual-stack scenario would look as follows

```
spec:
  apiVIP: 192.168.111.100
  ingressVIP: 192.168.111.101
  networking:
    machineNetwork:
      - cidr: "192.168.111.0/24"
      - cidr: "fd2e:6f44:5dd8:c956::/120"
    [...]
  provisionRequirements:
    controlPlaneAgents: 3
```

A sample CR presenting a remote worker node scenario would look as
follows

```
spec:
    apiVIP: 192.168.150.243
    ingressVIP: 192.168.150.242
    networking:
      clusterNetwork:
      - cidr: 10.128.0.0/14
        hostPrefix: 23
      machineNetwork:
      - cidr: 192.168.150.0/24
      - cidr: 192.168.122.0/24
      serviceNetwork:
      - 172.30.0.0/16
```

Contributes-to: MGMT-7315