MGMT-7315: Allow providing machine networks for multi-node non-DHCP
This commit adds an ability to manually provide machine networks in the
following scenario

* dual-stack OCP cluster
* non-DHCP mode
* VIPs manually provided

Previously whenever VIPs were provided, machine network was
automatically calculated based on them and the available networks on the
hosts. This is not a valid case anymore in case of a dual-stack cluster,
as multiple VIPs are not allowed.

In such a case we want to enable users to explicitly provide machine
networks so that the dual-stack cluster can be installed.

A sample CR presenting this scenario would look as follows

```
spec:
  apiVIP: 192.168.111.100
  ingressVIP: 192.168.111.101
  networking:
    machineNetwork:
      - cidr: "192.168.111.0/24"
      - cidr: "fd2e:6f44:5dd8:c956::/120"
    [...]
  provisionRequirements:
    controlPlaneAgents: 3
```

Contributes-to: MGMT-7315