MGMT-7768: Handle Host's Kind field during bind/unbind/register (#2578)
Host kind is set during Register operation. In case host already exists in DB,
the value stored in the DB is the one that is stored after register. Since host can be bound
to day2 cluster, unbound abd then bound again to day1 cluster we need to reset during unbind
operation. We also need to set the Kind based on the cluster, which is set during Register.
We must propagate this value into the host already stored on DB