Bug 1978691: Fix osImages reference passing (#2188)
Currently, when constructing OPENSHIFT_VERSIONS from the osImages on the
AgentServiceConfig, the values are passed by reference clumsily. This causes a
problem where the last in the `osImages` slice is used to determine the
root filesystem URL and ISO URL for all OPENSHIFT_VERSIONS.

If we use the reference directly from the agentserviceconfig, we should
avoid this problem.

Co-authored-by: David Zager <david.j.zager@gmail.com>