MGMT-4321: Extend network validators for dual-stack support (#2452)
This PR modifies the `ValidateIPAddressFamily` function so that it
restricts IPv6-only installations in non-IPv6 setups, but does not block
dual-stack.

The PR adds a validation whether machineNetwork, clusterNetwork, and
serviceNetwork are using IPv4 as the first subnet when deploying a
dual-stack cluster.

This PR refactors inventory.go by creating a wrapper function
ValidateIPAddresses over different network validations. This is in order
to reduce cyclomatic complexity.

The commit introduces also a helper DerefString function responsible for
converting an arbitrary argument to `*string` whenever possible. This is
useful because underlying types in `params.ClusterUpdateParams` and
`params.NewClusterParams` are not consistent (`string` vs `*string`).

Apart from that some unit tests are added to cover more validation
tests, mainly regarding network order and dual-stack support.

Contributes-to: MGMT-4321