Bug 2002495: Deploying OCP on RHOS with assisted-installer Failing validation blocks deployment (#2592)
The failed validation was: valid-platform.
The openstack platform is valid, the problem is with the network configuration.
Since openstack is the only platform in this validation and since it's
valid, the validation should be renamed:
`valid-platform-network-setting` and move to the network validations.
This way the user can update the netwok configuraion to resolve any
invalid settings in the network screen.