MGMT-7258: Set the imageURL to the image service if we're using it
This commit skips generating and saving the discovery image if a url to
the image service has been provided.

Additionally this commit ensures that we only reset the url if something
has changed by comparing the parameters from the url currently in the
infraenv with the new values provided with the request.