OCPBUGSM-34323: Use the hostname in the Spec if present (#2586)
The agent has two fields for the Hostname: Spec.Hostname and
Status.Inventory.Hostname.

The former is the one provided to the agent, the desired hostname. The
latter is the hostname discovered. The discovered hostname is sent
*before* the node's hostname is set. Relying solely on the discovered
hostname may cause inconsistencies when the desired hostname is
different from the discovered one.

Use the desired hostname (assuming the agent will succeed when setting
it) if present, otherwise rely on the discovered one.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>