MGMT-7282: Operator upgrade (#2577)
This PR adds new variables to be used by the operator upgrade CI job.

UPGRADE_OPERATOR - This is useful to determine which ASSISTED_OPENSHIFT_VERSION to be used. Before upgrading the operator, the value of this variable is true, which signals that the current operator needs to be updated. The value is derived from the default_ocp_versions.json. When set to true, ASSISTED_OPENSHIFT_VERSION gets the value 4.8. Otherwise, 4.9. After upgrading the operator, the value of this variable is false.

ASSISTED_STOP_AFTER_AGENT_DISCOVERY - This is useful when waiting for a full cluster deployment is not needed. For example, during an upgrade CI job, we may not need to wait for the full deployment in some of the phases of the job. Before upgrading the operator, the value of this variable is false, which preserves the existing behavior.
After upgrading the operator, the value of this variable is true.

ASSISTED_CLUSTER_NAME, ASSISTED_CLUSTER_DEPLOYMENT_NAME, ASSISTED_AGENT_CLUSTER_INSTALL_NAME, ASSISTED_INFRAENV_NAME - These are used for the upgraded cluster.

Consumed by: openshift/release#21539