MGMT-6763: Add support for dual-stack CI jobs
This commit implements handling `IP_STACK=v4v6` in the script used by
the CI to run dual-stack SNO and multi-node jobs.

We use `|length` in jinja template because of the behaviour of Ansible
env lookup that returns empty string if a variable is not defined and
effectively makes it impossible to distinguish between empty and
undefined variable.

Contributes-to: MGMT-6763