MGMT-7258: Ensure the url changes when necessary
Previously we would only update the url when some portion of the URL
itself changed. But in reality we need to force an update to the URL
even if something like ignition overrides or static network config
changes. This is because the URL change triggers the BMO to reboot the
host into the new iso. Without this, new changes to the iso wouldn't
make their way into previously started hosts.