MGMT-7847: Remove SUPPORT_L2 env var (#2650)
This environment variable was used to "enable L3" connectivity for day2
hosts. More precisely, this environment variable would disable
agent-side CIDR validation on the host, which was mostly relevant when
adding worker nodes.

The CIDR validation happens at the service level for *all* day1 hosts.
The validation for day2 hosts was pushed to the agent. In a PR pushed to
the agent repo[0] the logic to validate CIDR on the agent side is being
removed and this PR is focused on removing the variable controlling such
feature.

*IMPORTANT*
===

These 2 PRs will effectively remove any CIDR validation for day2 hosts
since the CIDR validation on the server side is only done for day1. This
is fine since the only thing we really care about is connectivity to the
VIP, which is still being validated by the agent.

CIDR validation doesn't add any value, considering that we are already
validating connectivity to the VIP.

[0] https://github.com/openshift/assisted-installer-agent/pull/235

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>