MGMT-7258: Use image service in assisted service (#2628)
* MGMT-7258: Set the imageURL to the image service if we're using it

This commit skips generating and saving the discovery image if a url to
the image service has been provided.

Additionally this commit ensures that we only reset the url if something
has changed by comparing the parameters from the url currently in the
infraenv with the new values provided with the request.

* MGMT-7258: Don't upload base images if we're using the image service

* MGMT-7258: Add the image service url to assisted service

The operator should deploy the image service and configure assisted
service to use to by default. If a user doesn't want this they should be
able to set the IMAGE_SERVICE_BASE_URL back to an empty string in the
override configmap.

* MGMT-7258: Ensure the url changes when necessary

Previously we would only update the url when some portion of the URL
itself changed. But in reality we need to force an update to the URL
even if something like ignition overrides or static network config
changes. This is because the URL change triggers the BMO to reboot the
host into the new iso. Without this, new changes to the iso wouldn't
make their way into previously started hosts.