MGMT-7687: Propagate Infrastructure Operator node selection and tolerations to its workloads (#2640)
* MGMT-7687: operator, node selection and tolerations

There is a common pattern in OpenShift to create infrastructure
machinesets [0]. These allow the cluster admins to have machines
specialized for running the things like Hive, ACM, assisted, etc.

The upcoming Open Cluster Management 2.4 release adds support for
running the operators in the infrastructure nodes. In order to satisfy
the expectation that the OCM components run in the infrastructure nodes,
we need to make sure that the workloads that we spawn remain in the same
node selection as the Infrastructure operator does.

The way this commit seeks to achieve that, is by retrieving the node
selection and the toleration that OCM set for the infrastructure
operator and set it for the image service and assisted service
deployments as well.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>

[0] https://docs.openshift.com/container-platform/4.8/machine_management/creating-infrastructure-machinesets.html

* MGMT-7687: Only propagate when running in k8s

In case a developer runs the operator locally with go run, we do not
want to fail because the retrieval of the operator Pod NodeSelector and
Tolerations fails. In those cases, we should just skip the propagation.

Signed-off-by: Antoni Segura Puimedon <antoni@redhat.com>

* MGMT-7687: Propagation only when POD_NAME specified

Instead of relying on the k8s service host, we can simply assess
`POD_NAME` environment variable not being specified as the indication
that we need to skip the node selection/tolerations.

Signed-off-by: David Zager <david.j.zager@gmail.com>

* MGMT-7687: Use non-caching client to get pod

Signed-off-by: David Zager <david.j.zager@gmail.com>

Co-authored-by: David Zager <david.j.zager@gmail.com>