Bug 2007515: Pass copy-network when infraEnv.StaticNetworkConfig is set
Originally, assisted-service would add the `--copy-network` argument to
the coreos-install command when the cluster.StaticNetworkConfigured flag
was set. This flag, however, is only been set to true for cloud-managed
deployments.

The cluster.StaticNetworkConfigured flag became a cached value of data
that should be considered dynamic, which broke the abstraction and the
separation of concerns of the various objects managed internally.
Specifically, it put in the cluster data that should only belong to the
infraenv.

This commit removes the use of `cluster.StaticNetworkConfigured` when
deciding to pass the `--copy-network` flag in favor of just checking
that the `infraenv.StaticNetworkConfig` is not empty.

The field in the model is not being removed in this commit as this will
be backported and it'd be preferable to not do database migrations in
backports. A follow-up commit will take care of removing the DB column.

Signed-off-by: Flavio Percoco <flavio@redhat.com>