Bug 2007539: Pass copy-network when infraEnv.StaticNetworkConfig is set (#2663)
* Bug 2007515: Pass copy-network when infraEnv.StaticNetworkConfig is set (#2662)

Originally, assisted-service would add the `--copy-network` argument to
the coreos-install command when the cluster.StaticNetworkConfigured flag
was set. This flag, however, is only been set to true for cloud-managed
deployments.

The cluster.StaticNetworkConfigured flag became a cached value of data
that should be considered dynamic, which broke the abstraction and the
separation of concerns of the various objects managed internally.
Specifically, it put in the cluster data that should only belong to the
infraenv.

This commit removes the use of `cluster.StaticNetworkConfigured` when
deciding to pass the `--copy-network` flag in favor of just checking
that the `infraenv.StaticNetworkConfig` is not empty.

The field in the model is not being removed in this commit as this will
be backported and it'd be preferable to not do database migrations in
backports. A follow-up commit will take care of removing the DB column.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>
(cherry picked from commit 8cf51023fecb0ef0d213b8d0d216176f903f0190)

* Bug 2007515: Fix regression introduced by copy-network bugfix (#2667)

Commit 8cf51023fecb0ef0d213b8d0d216176f903f0190 (#2662) introduced a bug
which broke late binding.

The `GetInfraEnvByClusterFromDB` function introduced by that commit
wrongly assumed that every infra env must have a corresponding cluster
ID. This is not the case for late-binding infra-envs where infra envs
can have no cluster ID.

This commit gets rid of the `GetInfraEnvByClusterFromDB` function and instead fetches
the infra-env database object during install command creation by taking the infraenv
ID from the host and using the existing `GetInfraEnvFromDB` function.

All "modern" hosts are guaranteed to have an infra env ID.

Older hosts in the database might not have an infra env ID, but we
do not expect any user to actually try to go through with "installation"
of these old hosts. In any case, the `constructHostInstallerArgs` has
been modified to handle a `nil` `*InfraEnv`.

Unit tests have been adjusted to use the correct infra env ID which is
actually set in the host, rather then re-generating a new infra env ID
when creating test-specific infra env objects.

(cherry picked from commit d3264105f8a3d6f7edfb3fbab269624e53cb7f74)

Co-authored-by: Omer Tuchfeld <otuchfel@redhat.com>