Bug 2001973: Fix handling nil value when setting disk encryption
This commit fixes handling of nil values when setting a feature usage
for disk encryption.

Currently it happens that when sending an update payload with
`disk_encryption: {}` we were trying to access fields inside the
payload, what was causing null pointer exceptions.

This commit also changes the way of determining the status of the
feature from looking at the cluster object into looking into the cluster
update object. Previously the check for nil values was taking into
account the current state of the cluster and not the desired state. With
this change, we are always comparing against the requested state from
the update payload.

Closes-bug: #2001973