Add Providers Support - enhancement implementation (#2546)
* OCPRHV-618: Base, Add the providers structure

This commit adds the providers directory.
It adds:
- Provider Interface: an interface that will define what needs to be implemented to be a supported provider.
- The Provider Registry - an API to interact with the providers, it will contain all the supported providers and contain functions to interact with a specific provider, or get information about supported providers.
- An initial implementation of the vsphere and baremetal providers.

* OCPRHV-619: Installcfg, Add AddProviderPlatformToInstallConfig

The AddProviderPlatformToInstallConfig is used to add a platform managed by a registered provider to an Install Config struct.
 That way each provider can implement its own specific parts.
 We also move existing functions/structs from installcfg.go to each provider.
 Only changes made from the original function is instead of recalculating the number of host we take the value from the cluster struct.
Also, we separate the installcfg builder into its own package - this is a general best practice and was needed so the implementation of the installcfg could import the provider registry and avoid a dependency cycle.
* dependency cycle: installcfg.go imported the providers (baremetal, vsphere and registry) and they imported it back, go doesn't allow it.

* OCPRHV-620: Inventory, Port updatePlatformSources and updatePlatformParams to SetPlatformDatabaseParameters and SetPlatformUsages

Here we port some of the inventory.go functions to the provider API.

* OCPRHV-621: Commons, Port GetClusterSupportedPlatforms to the provider registry

We want each provider to be able to implement its own logic regarding if a specific set of hosts is supported for installation.
This PR moves and refactors the logic on GetClusterSupportedPlatforms to be implemented by each provider.
Instead of checking for each host separately and then using Inner Join to combine the output we send each provider all the Discovered hosts and trust each provider to implement logic to check whether they are supported.

* OCPRHV-623: Port host validations to use ProviderRegistry

* OCPRHV-624: add mocks and fix some linting
- Add mock generation line for provider interfaces and generate mocks
- Linting - Fix RegisterClusterInternal cyclomatic complexity, we moved some of the easy validations to a separate function to reduce cyclomatic complexity
- minor lint changes

* OCPRHV-625: Unit tests, Inventory:
- add provider registry mock
- add getDefaultClusterCreateParams with platform set to baremetal
- Remove Update cluster tests:
The Update Cluster flow with updateClusterData can't be tested because the gomock doesn't supprot mocking a function that changes an argument that is not a pointer or a slice, see docs[1].
Since we can't change the updates map on the functions then the values will not be updated on the DB and the test will always fail.
- remove GetClusterSupportedPlatforms tests that are related to providers needs to be moved to the provider registry

[1] https://pkg.go.dev/github.com/golang/mock/gomock#Call.SetArg

* OCPRHV-626: Unit tests, installconfigbuilder, Fix install config builder tests

* OCPRHV-627: Unit tests, Add provider registry unit tests

* OCPRHV-628: Unit tests, fix hosts tests