Bug 2009412: when unbound host is bounded to day2 cluster, it stays in insufficient state (#2694)
The reason for insuffucient state is that although the cluster is day2, host kind is not set to be
day2 host.
Solution: during registration get bound cluster in case infra-env is bound or host is bound. Set
the host kind accordingly. Also, in the PostRegistration handler of transactions manager, in case host
already present in DB,update the kind field of the host based on the input host (input to the state machine).
This way we propagate the Kind change to the DB in case of re-registration (which is the flow in case of
bounded host)