NO-ISSUE: Fix typo and incorrect if condition added while improving operator upgrade job (#2703)
To be able to use community operators, the ASSISTED_DEPLOYMENT_METHOD should have the value as "from_community_operators". There was a typo and the value was set incorrectly to "community-operators". This resulted in calling the `setup_assisted_operator.sh` shell script with a wrong argument and the CI job for operator upgrade was failing with `exit 1`.