NO-ISSUE: Use correct catalog source name
This fixes and sets `catalog_source_name` correctly as `community-operators`. It was incorrectly getting set to `from_community_operators` and as a result, the `assisted-service-operator` subscription was not getting created.

Removed the use of `ASSISTED_DEPLOYMENT_METHOD` as it gets the value of `from_community_operators` and is used to see in the logs how the assisted operator is deployed, from community operators or from index image, for example.