Bug 2005440: Use current and new cluster data for dual-stack check
This commit extends the dual-stack validator to use both existing
cluster data and new cluster params when checking whether a
dual-stackness is requested.

The required behaviour comes from the fact that when using Kube API and
creating a new cluster, the initial payload contains empty Machine
Networks and only afterwards the reconcile loop is creating a cluster
update request with desired Machine Networks. This means that during
creation of the cluster we never have Cluster/Service Networks and
Machine Networks available in the same payload, so that a single
validation is not enough.

In order to overcome this issue, dual-stackness validation takes into
account the current payload as well as the previous state of the
cluster. This allows us to handle a scenario where Cluster/Service
Networks are updated separately from Machine Networks, even if all of
them were passed in the same CR.

Closes-bug: #2005440