Bug 2010479: Reduce open db connection limit to under the max (#2725)
The database is configured to accept a max of 100 connections. If we
also have our connection pool configured to 100 we'll hit the limit
before we stop giving out connections and error out some requests that
could have succeeded.

Reducing the max in the gorm connection pool will prevent these errors.

https://bugzilla.redhat.com/show_bug.cgi?id=2010479

Co-authored-by: Nick Carboni <ncarboni@redhat.com>