Bug 2013791: Delete host from DB if Agent creation fails
In Kube API flow, when an host registers a corresponding Agent CR is
created.
Before this patch, in case that the Agent CR creation fails, the host was still created
in the DB. That would lead to a state where the InfraEnv cannot be
deleted as it still has hosts connected.

To fix this issue, the host is deleted from DB if the creation of the
Agent CR fails