MGMT-7716: Add a template file for the image service (#2762)
* MGMT-7716: Add a template file for the image service

This will allow us to deploy the image service to the SaaS and will also allow
us to deploy new versions of the image service in the same way we do for the
assisted service.

Specifically this needs to be a separate template (rather than a new
deployment in the existing one) so that the app-sre automation can
handle automatically deploying it on changes to upstream repo.
If we used the same template we would need to manually update the tag
for the image service image we wanted to deploy in saas.yaml each time
it changed.

* MGMT-7716: Add new template to generate script

Also update the OS_IMAGES value to be up to date in the default
parameters in the template