MGMT-8021: update go-openapi/runtime
Specifically pick up the fix for
https://github.com/go-openapi/runtime/issues/158 which will allow us to
pass error response codes other than 403 out of the authorization code.

This is the minimum version jump to get the fix. Some later version
seemed to cause an issue with authz unit tests which will have to be
handled the next time we want to upgrade.