Revert "MGMT-7716: Add a template file for the image service (#2762)"
This reverts commit 18095889fd1de8516ac6f511095e90e77847afe6.

To properly track changes on the `main` branch this template needs to be
in the image service repo.

This will make updates to the os images a bit more difficult as we won't
be able to automate the update in a single place, but not having to
track the image service image tag separately is worth the effort.