MGMT-7923: Image service cloud authentication enhancement (#2733)
Authentication between the image service and assisted service when deployed
in the cloud requires a new strategy because the existing tokens used in the cloud
are too large and too short lived to be included in a pre-signed URL. Instead
of the existing authentication used in the cloud, assisted service will issue JWTs
using a symmetric key specific to each infraEnv. This token will be used
as authentication for requests to assisted service, specifically for the endpoints
required by the image service.