MGMT-8021: Move authz out of the authentication handler (#2780)
* MGMT-8021: update go-openapi/runtime

Specifically pick up the fix for
https://github.com/go-openapi/runtime/issues/158 which will allow us to
pass error response codes other than 403 out of the authorization code.

This is the minimum version jump to get the fix. Some later version
seemed to cause an issue with authz unit tests which will have to be
handled the next time we want to upgrade.

* MGMT-8021: make our error types satisfy the openapi error interface

According to the fix for https://github.com/go-openapi/runtime/issues/158
this will allow the message and code from our error types to be
propagated through the authorization middleware.

* MGMT-8021: move all authorization into the authz handler

* MGMT-8021: use the default api key authenticator

When we don't want any auth we still need to override this, but in all
the other cases we can use the default.