MGMT-7897: add image token auth implementation
Image auth is only implemented for rhsso and is accomplished in two
parts. Authentication parses the token and validates that the signature
is valid using the token for the infraEnv in the "sub" claim as well
as ensures that the token is not expired (using the "exp" claim).
Authorization then ensures that the infraEnv in the "sub" claim is the
same as the one in the request path.