Bug 2017444: Create BMH before the Machine (#2826)
The IPI Externally Provisioned workflow recommends creating the BMH before creating the Machine CR. This is
because when the Machine CR is created CAPBM will start reconciling it, which in our case will fail as the BMH
CR that referenced in the ProviderSpec won't exist.

This commit also fixes a bug by which the ConsumerRef in the BMH was missing the Machine's Kind and ApiVersion. This
is happening because the machine object that is used to get this data is instantiated by BMAC rather than coming from
a query. Instead of relying on such object, this commit switches to using the SchemaVersionGroup version and just hard-coding
the kind to Machine as we know it should not change

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>