Bug 2014081: Add user filter for /v2/events (#2798)
This change modifies /v2/events API endpoint query responses as follows:

Queries with no filters are not allowed:
The query response, in this case, would be gorm.ErrInvalidTransaction and the
API will return HTTP 400 StatusBadRequest. With and without a username in the
payload.

Filter resource IDs:
First, it will try to retrieve the given resource(s) from the DB.
- In case the resource does not exist (gorm.ErrRecordNotFound), the API will
  respond with HTTP 404 StatusNotFound.
- In case the DB returned some other error, the API will respond with
  HTTP 500 StatusInternalServerError.
- The reason for fetching the resources from the DB is the matching it needs
  to do between the username found in payload and resource owning username.
  DB resource query is always performed (user in the payload or not) to keep
  the API responses consistent in case of a DB error.

User matching:
- Done if user found in payload.
- Skipped if the user role is Admin

Implementation:
As stated above, the query now checks for the username in the payload.
For that, it imports the identity pkg, which causes a cyclic import.
The events implementation (that imports identity pkg) is now on a separate
pkg from the Interface to overcome this.