MGMT-8120: Change Events subsystem tests assertion
Recent CI results show that some tests may expect to get
two or three cluster events.

The reason is that it might take an extra second or two
to list events for a given cluster after it got registered.
So it will always get the following events:
- Registered cluster
- Successfully registered cluster

And sometimes it will also get:
- Updated status of the cluster to pending-for-input

This PR changes the tests to:
1. expect any number of events per cluster_id filtering
2. assert that all returned events do match the given cluster_id